[recovery] Package legacy recovery. Contributes to JB#25121
Initial recovery should work as it does on Jolla Phone where applicable.
To achieve that, this patch packages for Jolla Tablet the recovery from
Jolla Phone as it is today. This is not fully functional, but we
want to keep it this way to see the development diff more clearly.

Signed-off-by: Kalle Jokiniemi <kalle.jokiniemi@jolla.com>