[recovery] Bypass device lock in case of broken file system. Contributes to JB#30207
If user's file system is broken, there is no sense in asking for
device lock. In case of device theft, the thief will anyway need to
get access to file system to screw it up in order to bypass device
lock via this method. And if the thief has gotten access to file
system by other means, this device lock check does not help anyway.

Signed-off-by: Kalle Jokiniemi <kalle.jokiniemi@jolla.com>