[recovery] Don't unmount rootfs if there are other users. Contributes to JB#30614
Add is_single_user() function to check if we are a sole lock holder.
This allow to set-up LVM and mount device's rootfs on first lock and
unmount on last unlock.

This patch fixes situation when two telnet connections open to a device,
and on first logout rootfs was unmounted despite second connection still open.

Adjust copyright years and contact information in the scripts.

Signed-off-by: Igor Zhbanov <igor.zhbanov@jolla.com>

Conflicts:
	initrd/usr/bin/recovery-menu

Conflicts:
	initrd/usr/bin/recovery-menu