[initrd] Use explicit dot-file for FS resize status. Fixes JB#35480
On first boot LVM and filesystem resize can be interrupted in the middle
of the process by removing a battery. It could lead to a state when
LVM is resized but filesystems are not. And or init-scripts didn't try
to restart resizing.

By using dot-file as an indicator of successful filesystem resize
we can be sure that even after power loss resizing will continue
on the next boot.

Refactor messages to logging function.

Signed-off-by: Igor Zhbanov <igor.zhbanov@jolla.com>